requirements.txt
aiopg==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_postgres==1.0.6
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
distro==1.9.0
Django==5.1.1
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.1
geographiclib==2.0
geopy==2.4.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.0.8
mysqlclient==2.2.4
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.5
pillow==11.1.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rdkit==2025.3.3
redis==5.2.1
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.3.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.99
ultralytics-thop==2.0.14
urllib3==2.3.0
zope.interface==7.2
